backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# Redis
redis>=5.0.0

# HTTP client
httpx>=0.28.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Security
bcrypt>=4.2.0
pyjwt>=2.9.0

# Settings management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Scheduler
apscheduler>=3.10.0

# Development
python-multipart>=0.0.9
